requirements.txt
adversarial-robustness-toolbox==1.2.0
appnope==0.1.0
attrs==19.3.0
azure-common==1.1.25
azure-nspkg==3.0.2
azure-storage==0.36.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.1
blis==0.4.1
Bottleneck==1.3.2
captum==0.2.0
catalogue==1.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cox==0.1.post2
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
docopt==0.6.2
dpu-utils==0.2.9
entrypoints==0.3
fastai==1.0.60
fastprogress==0.2.2
gitdb==4.0.2
GitPython==3.1.0
grpcio==1.27.2
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==2.0.1
jupyterlab-server==1.0.7
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.0
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numexpr==2.7.1
numpy==1.18.1
nvidia-ml-py3==7.352.0
packaging==20.3
pandas==1.0.2
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
py3nvml==0.2.5
pycparser==2.20
Pygments==2.5.2
pynvx==1.0.0
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.7.1
QtPy==1.9.0
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
SetSimilaritySearch==0.1.7
six==1.14.0
smmap==3.0.1
soupsieve==2.0
spacy==2.2.4
srsly==1.0.2
tables==3.6.1
tensorboardX==2.0
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.8
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
xmltodict==0.12.0
zipp==2.2.0
